backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pypdf
google-genai
chromadb
python-dotenv
